PostController.java PostService.java 둘 다 ResponseEntity<ApiResponse>로 응답하도록 수정

+++ src/main/java/com/example/weboard/service/PostService.java
@@ -1,8 +1,11 @@
 package com.example.weboard.service;
 
+import com.example.weboard.dto.ApiResponse;
 import com.example.weboard.mapper.PostMapper;
 import com.example.weboard.dto.PostDTO;
 import lombok.RequiredArgsConstructor;
+import org.apache.coyote.BadRequestException;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -12,23 +15,44 @@ import java.util.List;
 
 public class PostService {
     private final PostMapper postMapper;
+    private final AuthService authService;
     public List<PostDTO> getPostAll() { return postMapper.getPostAll(); };
 
-    public List<PostDTO> getPostAllByOffset(Integer offset) { return postMapper.getPostAllByOffset(offset); };
-    public PostDTO getPostById(int postId){
-        return postMapper.getPostById(postId);
+    public ResponseEntity<ApiResponse> getPostAllByOffset(int offset) {
+        ApiResponse apiResponse = new ApiResponse(0, "성공적으로 게시물을 가져왔습니다.", postMapper.getPostAllByOffset(offset));
+        return ResponseEntity.status(200).body(apiResponse);
+    };
+    public ResponseEntity<ApiResponse> getPostById(int postId){
+        ApiResponse apiResponse = new ApiResponse(0, "성공적으로 게시물을 가져왔습니다.", postMapper.getPostById(postId));
+        return ResponseEntity.status(200).body(apiResponse);
     }
 
-    public void insertPost(PostDTO post){
+    public ResponseEntity<ApiResponse> insertPost(PostDTO post, String jwttoken){
+        Integer userId = authService.getIdFromToken(jwttoken);
+        post.setCreatedBy(userId);
         postMapper.insertPost(post);
+        ApiResponse apiResponse = new ApiResponse(0, "성공적으로 게시물을 작성했습니다.", null);
+        return ResponseEntity.status(201).body(apiResponse);
     }
 
-    public void updatePost(PostDTO post){
+    public ResponseEntity<ApiResponse> updatePost(PostDTO post, int postId, String jwttoken){
+        post.setPostId(postId);
+        Integer userId = authService.getIdFromToken(jwttoken);
+        post.setUpdatedBy(userId);
         postMapper.updatePost(post);
+        ApiResponse apiResponse = new ApiResponse(0, "성공적으로 게시물을 수정했습니다.", null);
+        return ResponseEntity.status(200).body(apiResponse);
     }
 
-    public void deletePost(int postId){
+    public ResponseEntity<ApiResponse> deletePost(int postId, String jwttoken) throws BadRequestException {
+        int createdById = postMapper.getPostById(postId).getCreatedBy();
+        int idFromJwt = authService.getIdFromToken(jwttoken);
+        if(idFromJwt!=createdById){
+            throw new BadRequestException("타인의 댓글은 삭제할 수 없습니다.");
+        }
         postMapper.deletePost(postId);
+        ApiResponse apiResponse = new ApiResponse(0, "성공적으로 게시물을 삭제했습니다.", null);
+        return ResponseEntity.status(200).body(apiResponse);
     }
 
 }
